% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{Alcohol warehouse and retail sales}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1000 rows and 9 columns.
}
\source{
https://catalog.data.gov/dataset/warehouse-and-retail-sales
}
\usage{
alcohol
}
\description{
\describe{
\item{year}{Year}
\item{month}{Month}
\item{supplier}{Supplier}
\item{item_code}{Item code}
\item{item_description}{Item description}
\item{item_type}{Item type}
\item{retail_sales}{Retail sales}
\item{retail_transfers}{Retail transfers}
\item{warehouse_sales}{Warehouse sales}
}
}
\keyword{datasets}
