% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{args_model}
\alias{args_model}
\title{Specify the default arguments of the model}
\usage{
args_model(vendor)
}
\arguments{
\item{vendor}{The vendor of the model.}
}
\description{
Specify the default arguments of the model
}
\seealso{
\code{\link[=args_ollama]{args_ollama()}}
}
