% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{as_prompt}
\alias{as_prompt}
\title{Make it into prompt}
\usage{
as_prompt(x, role, .content_envir = rlang::caller_env())
}
\arguments{
\item{x}{An input.}

\item{role}{The role.}
}
\description{
Make it into prompt
}
\seealso{
\code{\link[=is_prompt]{is_prompt()}}
}
