% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{emend_assist}
\alias{emend_assist}
\title{emend the general assistant}
\usage{
emend_assist(
  prompt = NULL,
  model = emend_get_option("model"),
  format = model$args$format,
  .content_envir = rlang::caller_env(),
  ...
)
}
\arguments{
\item{prompt}{A list of prompts or text with a prompt.}

\item{model}{A list of model and vendor name. Use model helpers like \code{model_ollama()} and \code{model_openai()}.}
}
\description{
emend the general assistant
}
