% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{emend_describe_image}
\alias{emend_describe_image}
\title{Describe the input image}
\usage{
emend_describe_image(x, model = emend_get_option("model"))
}
\arguments{
\item{x}{A url or path to the image.}

\item{model}{The LLM to use.}
}
\description{
This requires a multimodal LLM like \verb{llava:7b} to be installed.
}
