% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{emend_lvl_match}
\alias{emend_lvl_match}
\alias{emend_fct_match}
\title{Match the input factor to supplied levels}
\usage{
emend_lvl_match(.f, levels = NULL, ...)

emend_fct_match(.f, levels = NULL, ...)
}
\arguments{
\item{.f}{A factor.}

\item{levels}{The levels of the factor.}

\item{...}{Other prompts to the LLM.}
}
\description{
Match the input factor to supplied levels
}
\seealso{
\code{\link[=emend_lvl_sweep]{emend_lvl_sweep()}}
}
