% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{emend_option}
\alias{emend_option}
\alias{emend_get_option}
\alias{emend_set_option}
\alias{emend_set_model}
\alias{emend_set_model_args}
\alias{emend_set_language}
\title{Get or set an option for the \code{emend} package}
\usage{
emend_get_option(x = c("model", "language", "all"))

emend_set_option(x = c("model", "language"), val)

emend_set_model(val)

emend_set_model_args(...)

emend_set_language(val)
}
\arguments{
\item{x}{The name of the option to get or set.}

\item{val}{The value of the option to set.}
}
\description{
Get or set an option for the \code{emend} package
}
