% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{emend_what_language}
\alias{emend_what_language}
\title{Identify the language in the text.}
\usage{
emend_what_language(text, chat = NULL)
}
\arguments{
\item{text}{A string or a factor that contains text information.}

\item{chat}{A chat object defined by ellmer}
}
\description{
Identify the language in the text.
}
\examples{
emend_what_language(c("猿も木から落ちる", "你好", "bon appetit")

}
