% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/datasets.R
\docType{data}
\name{hotel}
\alias{hotel}
\title{Hotel reviews}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13193 rows and 8 columns.

A data.frame with 13,193 rows and 8 columns
\describe{
\item{reviewer_name}{The name of the reviewer.}
\item{reviewer_nationality}{The nationality of the reviewer.}
\item{reviewer_rating}{Reviewer's rating.}
\item{review_date}{Date of the review.}
\item{review_title}{Title of the review.}
\item{review_text}{Text of the review.}
\item{hotel_name}{Name of the hotel being reviewed.}
\item{avg_rating}{Average rating of the hotel.}
}
}
\source{
https://www.kaggle.com/datasets/nikitaryabukhin/reviewshotel
}
\usage{
hotel

hotel
}
\description{
A sample review of hotels.
}
\keyword{datasets}
