% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{is_prompt}
\alias{is_prompt}
\alias{is_prompt_user}
\alias{is_prompt_system}
\alias{is_prompt_assistant}
\title{Is this a prompt?}
\usage{
is_prompt(x, role)

is_prompt_user(x)

is_prompt_system(x)

is_prompt_assistant(x)
}
\arguments{
\item{x}{An input.}
}
\value{
A logical value to indicate if it is a prompt or not.
}
\description{
Is this a prompt?
}
\examples{
is_prompt(prompt_user("Hi!"))
}
\seealso{
\code{\link[=as_prompt]{as_prompt()}}
}
