% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{likerts}
\alias{likerts}
\title{A collection of different likert scales}
\format{
A data.frame with 40 rows and 9 columns
\describe{
\item{likert1}{A 7-point agreeableness likert scale.}
\item{likert2}{A 5-point agreeableness likert scale.}
\item{likert3}{A 5-point agreeableness likert scale as a sentence.}
\item{likert4}{A 5-point frequency likert scale.}
\item{likert5}{A 5-point rating likert scale.}
\item{likert6}{A 5-point likelihood likert scale.}
\item{likert7}{A 5-point likert scale.}
\item{likert8}{A 5-point satisfaction likert scale.}
\item{likert9}{A 6-point priority likert scale.}
}
}
\usage{
likerts
}
\description{
A data set containing 9 different likert scales.
}
\keyword{datasets}
