% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_emend.R
\name{lvl_match}
\alias{lvl_match}
\alias{fct_match}
\title{Match the input factor to supplied levels.}
\usage{
lvl_match(.f, levels = NULL, model = NULL, ...)

fct_match(.f, levels = NULL, model = NULL, ...)
}
\arguments{
\item{.f}{A factor.}

\item{levels}{The levels of the factor.}

\item{model}{The model defined by \code{\link[=set_model]{set_model()}}.}

\item{...}{Other prompts to the LLM.}
}
\description{
Match the input factor to supplied levels.
}
\seealso{
\code{\link[=lvl_sweep]{lvl_sweep()}}
}
