% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_emend.R
\name{fct_reorder}
\alias{fct_reorder}
\alias{lvl_order}
\title{Reorder levels of an ordinal factor}
\usage{
fct_reorder(.f, model = NULL, ...)

lvl_order(.f, model = NULL, copy = FALSE, ...)
}
\arguments{
\item{.f}{A character vector that is assumed to be an ordinal factor.}

\item{model}{A model defined by \code{\link[=set_model]{set_model()}}.}

\item{...}{Extra prompts to the LLM.}

\item{copy}{A logical value to indicate whether the output should be copied into the
user's clipboard.}
}
\description{
This function reorders the levels of a factor based on the sentiment scores of the levels.
Using this function can be expensive (depending on the LLM used and user's computer spec)
so users may wish to use this interactively only and copy the output into their script.
}
\examples{
# to get the new level order
# users should check if the new order
lvl_order(likerts$likert1)
# `copy = TRUE` copies the output into clipboard in a format that can be
# entered easiliy in the user's script
lvl_order(likerts$likert1, copy = TRUE)
# to apply the new levels directly to the input
fct_reorder(likerts$likert1)

}
