% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{lvl_sweep}
\alias{lvl_sweep}
\title{Clean up the levels for the input factor.}
\usage{
lvl_sweep(.f, model = NULL, ...)
}
\arguments{
\item{.f}{A factor.}

\item{model}{The model defined by set_model.}

\item{...}{Other prompts to the LLM.}
}
\description{
Clean up the levels for the input factor.
}
