% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{messy}
\alias{messy}
\title{A collection of messy inputs}
\format{
A list of 3 character vectors
\describe{
\item{country}{A character vector of countries.}
\item{suburb}{A character vector of suburbs in Australia with various typos.}
\item{school}{A character vector of schools or college (with typos) at the Australian National University.}
}
}
\usage{
messy
}
\description{
A synthetic dataset that contains inputs with some common standardisation
issues.
}
\keyword{datasets}
