% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mistral_model_list}
\alias{mistral_model_list}
\title{Get the list of Mistral AI models}
\usage{
mistral_model_list(api = NULL)
}
\arguments{
\item{api}{The Mistral AI API key.}
}
\description{
Get the list of Mistral AI models
}
