% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_mistral}
\alias{model_mistral}
\title{Set a model from Mistral AI}
\usage{
model_mistral(
  model = "open-mistral-7b",
  temperature = 0,
  top_p = 0.1,
  max_tokens = 100,
  min_tokens = 0,
  stop = NULL,
  seed = NA,
  format = "none",
  safe_prompt = FALSE,
  tools = NULL,
  tool_choice = "auto"
)
}
\arguments{
\item{temperature}{The temperature of the model where higher temperature means higher creativity.}

\item{top_p}{The nucleus sampling or penalty. It limits the cumulative probability of the most likely tokens.
Higher values allow more tokens and diverse responses, and while lower values are more focused and constrained answers.}

\item{max_tokens}{The maximum number of tokens to generate. This limits the length of the response.
Tokens refers to a unit of text the model reads and can vary from one character to several words,
varying based on the model. As a rule of thumb 1 token is approximately 4 characters or 0.75 words
for English text.}

\item{min_tokens}{The minimum number of tokens to generate.}

\item{stop}{A list of stop sequences to use.}

\item{format}{The format}

\item{safe_prompt}{Whether to use safe prompt or not.}

\item{tools}{The list of tools. Not implemented.}

\item{tool_choice}{The choice of tool. Not implemented.}
}
\description{
Set a model from Mistral AI
}
\seealso{
Other large-lanugage-models: 
\code{\link{model_ollama}()},
\code{\link{model_openai}()},
\code{\link{model_vendor}()}
}
\concept{large-lanugage-models}
