% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{model_vendor}
\alias{model_vendor}
\alias{model_ollama}
\alias{model_openai}
\title{Create a list of model and vendor}
\usage{
model_vendor(model, vendor)

model_ollama(model = "llama3.1:8b")

model_openai(model = "gpt-4o-mini")
}
\arguments{
\item{model}{The name of the model}

\item{vendor}{The name of the vendor}
}
\description{
Create a list of model and vendor
}
