% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{ollama_model_list}
\alias{ollama_model_list}
\title{Get the list of ollama models available locally}
\usage{
ollama_model_list(port = sai_get_option("port"))
}
\arguments{
\item{port}{The port number.}
}
\description{
Get the list of ollama models available locally
}
