% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{openai_model_list}
\alias{openai_model_list}
\title{Get the list of OpenAI models}
\usage{
openai_model_list(api = NULL)
}
\arguments{
\item{api}{The OpenAI API key. Provide "none" to get the list without an API key.}
}
\description{
Get the list of OpenAI models
}
