% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{prompt_user}
\alias{prompt_user}
\alias{prompt_assistant}
\alias{prompt_system}
\title{Prompt input for the LLM}
\usage{
prompt_user(
  content,
  images = NULL,
  tool_calls = NULL,
  .content_envir = rlang::caller_env(),
  .collapse_regex = "[*]$"
)

prompt_assistant(
  content,
  images = NULL,
  tool_calls = NULL,
  .content_envir = rlang::caller_env(),
  .collapse_regex = "[*]$"
)

prompt_system(
  content,
  images = NULL,
  tool_calls = NULL,
  .content_envir = rlang::caller_env(),
  .collapse_regex = "[*]$"
)
}
\arguments{
\item{content}{A single text with any content in "{}" evaluated as an R expression.
If the bracket contains "\emph{" at the end (or regular expression match as specified in \code{.collapse_regex})
then it collapses into a comma separated list in a square bracket with each element single quoted.
E.g. "{1:3}}" expands to \link{'1', '2', '3'}.}

\item{images}{A list of path or url to the images (or base64 encoded images).}

\item{tool_calls}{Currently doesn't work as intended.}

\item{.content_envir}{The environment used to evaluate \code{content}.}

\item{.collapse_regex}{The regular expression match to collapse the content.}

\item{role}{The role should be "user", "assistant" or "system".}
}
\description{
A helper function to create the prompt input.
}
\examples{
prompt_user("What should I eat today? Pick one of {c('Ramen', 'Pizza', 'Burger')*}.")
}
