% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{recipes}
\alias{recipes}
\title{Recipes}
\format{
A data.frame with 200 rows and 7 columns.
\describe{
\item{name}{The name of the dish.}
\item{ingredients}{The list of ingredients.}
\item{url}{The URL of the recipe.}
\item{image}{An image of the dish.}
\item{cook_time}{Cooking time.}
\item{prep_time}{Preperation time.}
\item{servings}{The number of servings in text.}
}
}
\source{
https://github.com/jakevdp/open-recipe-data/tree/main
}
\usage{
recipes
}
\description{
A sample of 200 recipes.
}
\keyword{datasets}
