% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{restaurant}
\alias{restaurant}
\title{Review of Iori restaurant}
\format{
A data.frame with 20 rows and 2 columns
\describe{
\item{review}{The text of the review.}
\item{rating}{A rating out of 5.}
}
}
\usage{
restaurant
}
\description{
Iori is a Japanese restaurant in Canberra, Australia.
The data contains a sample of 20 reviews from Google maps.
}
\keyword{datasets}
