% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAI-package.R
\docType{package}
\name{SAI-package}
\alias{SAI}
\alias{SAI-package}
\title{SAI: Standarise data with an AI assistant}
\description{
The package provides a simple interface to use the large language models to clean data.
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]

}
\keyword{internal}
