% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{sai_assist}
\alias{sai_assist}
\title{SAI the general assistant}
\usage{
sai_assist(
  prompt = NULL,
  model = sai_get_option("model"),
  args = args_model(vendor),
  .content_envir = rlang::caller_env(),
  ...
)
}
\arguments{
\item{prompt}{A list of prompts or text with a prompt.}

\item{model}{A list of model and vendor name. Use model helpers like \code{model_ollama()} and \code{model_openai()}.}

\item{args}{The arguments for the \code{model}.}
}
\description{
SAI the general assistant
}
