% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{sai_clean_address}
\alias{sai_clean_address}
\title{Standardise address format}
\usage{
sai_clean_address(
  address_vector,
  output_format = "155/255 Dickson St, Acton ACT 2602",
  copy = FALSE,
  ...
)
}
\arguments{
\item{address_vector}{A character vector that is assumed to be addresses.}

\item{output_format}{A character value to specify output address format.}
}
\description{
This function standardise inconsistent address formats.
}
