% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{sai_clean_date}
\alias{sai_clean_date}
\title{Standardise date format}
\usage{
sai_clean_date(
  dates_vector,
  input_format = "DD/MM/YYYY",
  output_format = "YYYY-MM-DD",
  copy = FALSE,
  ...
)
}
\arguments{
\item{dates_vector}{A character vector that is assumed to be dates.}

\item{input_format}{A character value to specify input date format.}

\item{output_format}{A character value to specify output date format.}
}
\description{
This function standardise inconsistent date formats.
}
