% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sai.R
\name{sai_coxph}
\alias{sai_coxph}
\title{SAI for Proportional Hazards Regression Model}
\usage{
sai_coxph(data, y, alpha = 0.05, mustin = NULL)
}
\arguments{
\item{data}{the data.frame which contain responce variable and explanatory variables.}

\item{y}{the responce variable "name", is a character.}
}
\value{
\bold{form} is the formulas

\bold{pvalue} is the p.value correspond to the formulas in \bold{form}

\bold{sig} is value 1 or 0, represents that if any \bold{pvalue} is significant in the model.
}
\description{
Find all combinations of explanatory variables, and show the significant linear regression model.
}
\examples{
sai_coxph(data,y, alpha = 0.05)
}
