% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{sai_lvl_match}
\alias{sai_lvl_match}
\alias{sai_fct_match}
\title{Match the input factor to supplied levels}
\usage{
sai_lvl_match(.f, levels = NULL, ...)

sai_fct_match(.f, levels = NULL, ...)
}
\arguments{
\item{.f}{A factor.}

\item{levels}{The levels of the factor.}

\item{...}{Other prompts to the LLM.}
}
\description{
Match the input factor to supplied levels
}
\seealso{
\code{\link[=sai_lvl_sweep]{sai_lvl_sweep()}}
}
