% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{sai_option}
\alias{sai_option}
\alias{sai_get_option}
\alias{sai_set_option}
\alias{sai_set_model}
\alias{sai_set_port}
\alias{sai_set_language}
\title{Get or set an option for the \code{sai} package}
\usage{
sai_get_option(
  x = c("model", "vendor", "port", "language", "num_predict", "all")
)

sai_set_option(x = c("model", "port", "num_predict", "language"), val)

sai_set_model(val)

sai_set_port(val)

sai_set_language(val)
}
\arguments{
\item{x}{The name of the option to get or set.}

\item{val}{The value of the option to set.}
}
\description{
Get or set an option for the \code{sai} package
}
