% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sai.R
\name{sai_screen}
\alias{sai_screen}
\title{SAI for SCREEN}
\usage{
sai_screen(data, y, model)
}
\arguments{
\item{data}{the data.frame which contain responce variable and explanatory variables.}

\item{y}{the responce variable "name", is a character.}

\item{model}{the model's name, is a character, `linear` for linear regression, `logit` for logistic regression, `poisson` for poisson regression, `ordinal_logistic` for ordinal logistic regression}
}
\value{
it will output a data.frame with univariate regreesion model to all explanatory variables, the p.value \bold{will be sorted} from smallest to lagest. So that we can easily find out which variable will effect the model dramatically.
}
\description{
do \bold{univariate} regression to all explanatory variables.
}
\examples{
sai_screen(data = iris, y = "Sepal.Length", model = "linear")
}
