% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{sai_translate}
\alias{sai_translate}
\title{Translate text from one language to another}
\usage{
sai_translate(text, from = NULL, to = sai_get_option("language"))
}
\arguments{
\item{text}{The text to translate.}

\item{from}{The language to translate from. The default is NULL (not specified).}

\item{to}{The language to translate to. The default is "English". The default
can be modified using \code{sai_set_language()}.}
}
\description{
Translate text from one language to another
}
\examples{
# example code
sai_translate(c("猿も木から落ちる", "在公寓区，周边有很多吃的还有大型超市，很方便"))

}
\seealso{
Other translate: 
\code{\link{sai_is_language}()},
\code{\link{sai_what_language}()}
}
\concept{translate}
