% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{salary}
\alias{salary}
\title{Ask A Manager Salary Survey 2021}
\format{
A data.frame with 28,083 rows and 18 columns
\describe{
\item{timestamp}{The timestampe of response.}
\item{age}{Age cateogry. "How old are you?"}
\item{industry}{Categorical but respondents can enter text for Other. "What industry do you work in?"}
\item{job_title}{Text entry. "Job title"}
\item{job_title_context}{Text entry. "If your job title needs additional context, please clarify here:"}
\item{salary_annual}{Text entry. "What is your annual salary? (You'll indicate the currency in a later question. If you are part-time or hourly, please enter an annualized equivalent -- what you would earn if you worked the job 40 hours a week, 52 weeks a year.)"}
\item{salarly_additional}{Text entry. "How much additional monetary compensation do you get, if any (for example, bonuses or overtime in an average year)? Please only include monetary compensation here, not the value of benefits."}
\item{currency}{Categorical entry. "Please indicate the currency"}
\item{currency_other}{Text entry. "If "Other," please indicate the currency here:"}
\item{salary_context}{Text entry. "If your income needs additional context, please provide it here:"}
\item{country}{Text entry. "What country do you work in?"}
\item{state}{Categorical entry. "If you're in the U.S., what state do you work in?"}
\item{city}{Text entry. "What city do you work in?"}
\item{experience_overall}{Categorical entry. "How many years of professional work experience do you have overall?"}
\item{experience_in_field}{Categorical entry. "How many years of professional work experience do you have in your field?"}
\item{education}{Categorical entry."What is your highest level of education completed?"}
\item{gender}{Categorical entry. "What is your gender?"}
\item{race}{Multiselect entry. "What is your race? (Choose all that apply.)"}
}
}
\source{
https://www.askamanager.org/2021/04/how-much-money-do-you-make-4.html
}
\usage{
salary
}
\description{
Ask A Manager Salary Survey 2021
}
\keyword{datasets}
