% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{set_model}
\alias{set_model}
\title{Define the model you are going to use for the analyses.}
\usage{
set_model(provider = NULL, model = NULL)
}
\arguments{
\item{provider}{The provider name. Choose one from {"ollama", "openai"}.}

\item{model}{The model name, e.g. "llama3.1:8b" for Ollama and "gpt-4o-mini" for OpenAI.}
}
\description{
Define the model you are going to use for the analyses.
}
