% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/datasets.R
\docType{data}
\name{consumer}
\alias{consumer}
\title{Amazon consumer reviews}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4549 rows and 9 columns.

A data.frame with 4549 rows and 11 columns
\describe{
\item{overall}{Overall rating of the product.}
\item{verified}{Whether the reviewer is verified or not.}
\item{review_date}{Review date.}
\item{review_id}{A unique identifier of the reviewer.}
\item{reviewer_name}{The name of the reviewer.}
\item{review_text}{The text of the review.}
\item{summary}{Summary of the review.}
\item{vote_helpful}{The number of helpful votes of the review.}
\item{image}{The images that the reviewer post after they receive the product.}
}
}
\source{
Jianmo Ni, Jiacheng Li, Julian McAuley (2019) Justifying recommendations using distantly-labeled reviews and fined-grained aspects. Empirical Methods in Natural Language Processing .
}
\usage{
consumer

consumer
}
\description{
A sample of reviews from Amazon in the applicances category.
}
\keyword{datasets}
