% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{emend_fct_reorder}
\alias{emend_fct_reorder}
\title{Reorder the levels of the input factor in a meaningful way.}
\usage{
emend_fct_reorder(.f, chat = NULL, ...)
}
\arguments{
\item{.f}{A factor.}

\item{chat}{A chat object defined by ellmer}

\item{...}{Other prompts to the LLM.}
}
\description{
Reorder the levels of the input factor in a meaningful way.
}
\examples{
chat <- chat_ollama(model = "llama3.1:8b", seed = 0, echo = "none")
emend_fct_reorder(likerts$likert1, chat = chat) |> levels()

}
