% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{emend_translate}
\alias{emend_translate}
\title{Translate text from one language to another.}
\usage{
emend_translate(text, from = NULL, to = "English", chat = NULL)
}
\arguments{
\item{text}{The text to translate.}

\item{from}{The language to translate from. The default is NULL (not specified).}

\item{to}{The language to translate to. The default is "English".}
}
\description{
Translate text from one language to another.
}
\examples{
emend_translate(c("猿も木から落ちる", "你好", "bon appetit"))

}
