% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{emend_what_language}
\alias{emend_what_language}
\title{Determine the language of the input text}
\usage{
emend_what_language(text)
}
\arguments{
\item{text}{The text.}
}
\description{
Determine the language of the input text
}
\examples{
emend_what_language(c("猿も木から落ちる", "你好", "bon appetit"))
}
\seealso{
Other translate: 
\code{\link{emend_is_language}()},
\code{\link{emend_translate}()}
}
\concept{translate}
