% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_vendor}
\alias{model_vendor}
\title{Create a list of model and vendor}
\usage{
model_vendor(model, vendor, ...)
}
\arguments{
\item{model}{The name of the model}

\item{vendor}{The name of the vendor}
}
\description{
Create a list of model and vendor
}
\seealso{
Other large-lanugage-models: 
\code{\link{model_mistral}()},
\code{\link{model_ollama}()},
\code{\link{model_openai}()}
}
\concept{large-lanugage-models}
