% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist.R
\name{sai_assist}
\alias{sai_assist}
\title{SAI the general assistant}
\usage{
sai_assist(
  prompt = NULL,
  vendor = sai_get_option("vendor"),
  model = sai_get_option("model"),
  port = sai_get_option("port"),
  format = c("none", "json"),
  tools = NULL,
  temperature = 0,
  stop = NA,
  seed = NA,
  top_k = 1,
  top_p = 0.1,
  num_predict = sai_get_option("num_predict"),
  keep_alive = "5m",
  .content_envir = rlang::caller_env(),
  ...
)
}
\arguments{
\item{prompt}{A list of prompts or text with a prompt.}

\item{vendor}{The vendor name. By default is ollama.}

\item{model}{The ollama model name, e.g. \verb{llama3.2:1b} (default), \verb{gemma:7b}, \verb{llama3.1:8b}}

\item{port}{The port number.}

\item{format}{The format to use. Default is NA and currently only other option is "json".}

\item{temperature}{The temperature of the model where higher temperature means higher creativity.}

\item{stop}{A list of stop sequences to use.}
}
\description{
SAI the general assistant
}
